Another correction to test_cli

+++ humanfriendly/tests.py
@@ -674,7 +674,7 @@ class HumanFriendlyTestCase(TestCase):
         returncode, output = run_cli(main, '--parse-length=5 km')
         assert int(output) == humanfriendly.parse_length('5 km')
         returncode, output = run_cli(main, '--parse-length=1.05 mm')
-        assert int(output) == humanfriendly.parse_length('1.05 mm')
+        assert float(output) == humanfriendly.parse_length('1.05 mm')
         # Test `humanfriendly --run-command'.
         returncode, output = run_cli(main, '--run-command', 'bash', '-c', 'sleep 2 && exit 42')
         assert returncode == 42
